FEATURE: use category icons and emojis from core (#85)
Adds support for core changes to category icons and emojis, allowing them to be shown on category banners.

We will soon be deprecating the Category Icons theme component since the functionality has been moved into core.

For categories that don't have an icon or emoji (ie. square badge), we just show the category name when the theme has the show icons setting enabled.